# Even faster array parsing