--- a/index.js
+++ b/index.js
@@ -20,6 +20,14 @@ const NULL_STRING = 'NULL'
 function makeParseArrayWithTransform (transform) {
   const haveTransform = transform != null
   return function parseArray (str) {
+    const rbraceIndex = str.length - 1
+    if (rbraceIndex === 1) {
+      return []
+    }
+    if (str[rbraceIndex] !== RBRACE) {
+      throw new Error('Invalid array text - must end with }')
+    }
+
     // If starts with `[`, it is specifying the index boundas. Skip past first `=`.
     let position = 0
     if (str[position] === LBRACKET) {
@@ -29,17 +37,12 @@ function makeParseArrayWithTransform (transform) {
     if (str[position++] !== LBRACE) {
       throw new Error('Invalid array text - must start with {')
     }
-    const rbraceIndex = str.length - 1
-    if (str[rbraceIndex] !== RBRACE) {
-      throw new Error('Invalid array text - must end with }')
-    }
     const output = []
     let current = output
     const stack = []
 
     let currentStringStart = position
-    const currentStringParts = []
-    let hasStringParts = false
+    let currentString = null
     let expectValue = true
 
     for (; position < rbraceIndex; ++position) {
@@ -56,8 +59,11 @@ function makeParseArrayWithTransform (transform) {
         while (backSlash !== -1 && backSlash < dquot) {
           position = backSlash
           const part = str.slice(currentStringStart, position)
-          currentStringParts.push(part)
-          hasStringParts = true
+          if (currentString === null) {
+            currentString = part
+          } else {
+            currentString += part
+          }
           currentStringStart = ++position
           if (dquot === position++) {
             // This was an escaped doublequote; find the next one!
@@ -68,13 +74,12 @@ function makeParseArrayWithTransform (transform) {
         }
         position = dquot
         const part = str.slice(currentStringStart, position)
-        if (hasStringParts) {
-          const final = currentStringParts.join('') + part
-          current.push(haveTransform ? transform(final) : final)
-          currentStringParts.length = 0
-          hasStringParts = false
-        } else {
+        if (currentString === null) {
           current.push(haveTransform ? transform(part) : part)
+        } else {
+          currentString += part
+          current.push(haveTransform ? transform(currentString) : currentString)
+          currentString = null
         }
         expectValue = false
       } else if (char === LBRACE) {